100-get_endianness.c

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -9,5 +9,5 @@ int get_endianness(void)
 	unsigned int i = 1;
 	char *c = (char *) &i;
 
-	return (*c);
+	return ((int)*c);
 }
